working with administration

+++ paynament/models.py
@@ -3,4 +3,4 @@ from accounts.models import User
 
 class Balance(models.Model):
     user = models.OneToOneField(User, related_name='balance', on_delete=models.CASCADE, verbose_name='Клиент')
-    balance = models.FloatField(default=0.00, verbose_name='Баланс')
+    balance = models.DecimalField(decimal_places=2, max_digits=8,default=0.00, verbose_name='Баланс')
